requirements.txt
absl-py==2.0.0
asgiref>=3.8.1
astunparse==1.6.3
blinker==1.8.2
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap5==2025.6
gunicorn==21.2.0
cryptography==38.0.4
cssselect==1.3.0
cycler==0.12.1
DataRecorder==3.6.2
decorator==5.2.1
Django==4.2.25
django-allauth==65.14.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-htmx==1.26.0
djangorestframework==3.16.1
dm-tree==0.1.8
dnspython==2.6.1
DownloadKit==2.0.7
DrissionPage==4.1.1.2
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==2.2.5
Flask-Cors==4.0.1
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2024.5.0
future==1.0.0
gast==0.5.4
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.25.2
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphviz==0.20.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httplib2==0.31.2
huggingface-hub==0.23.0
hyperframe==6.1.0
idna==3.6
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
kaitaistruct==0.10
kiwisolver==1.4.5
ldap3==2.9.1
libclang==16.0.6
lxml==6.0.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.8
namex==0.0.7
networkx==3.2.1
nltk==3.8.1
numpy
oauthlib==3.2.2
openpyxl==3.1.5
packaging==23.2
pandas
passlib==1.7.4
Pillow==10.1.0
proglog==0.1.12
proto-plus==1.27.1
protobuf==4.23.4
psutil==7.1.2
psycopg2-binary==2.9.11
publicsuffix2==2.20191221
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.23
pydivert==2.1.0
Pygments==2.17.2
PyJWT==2.11.0
pymongo==4.7.2
pyOpenSSL==22.1.0
pyparsing==3.3.2
pyperclip==1.8.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.2.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==8.2
regex==2024.5.15
reportlab==4.4.4
requests==2.31.0
requests-file==3.0.1
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.14
safetensors==0.4.3
segno==1.6.6
setuptools==75.8.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.2.0
tldextract==5.3.0
tokenizers==0.19.1
tornado==6.5.2
tqdm==4.66.4
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.2.0
urllib3==2.1.0
urwid==2.1.2
websocket-client==1.9.0
Werkzeug==3.0.1
whitenoise==6.11.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
zstandard==0.19.0
